examples/glan_replication/requirements.txt
openai
python-dotenv
# pandas (will be useful for data handling later)
# numpy
# scikit-learn (for potential embedding/similarity checks in decontamination)
